refactor out session and impression classes
